set TIMEFORMAT to empty (closes #7)
Only GNU time supports the "-f" switch to change the output format of `time`. We
decide to not use a special format as default because GNU `time` is not
available on Mac.
The highlighting of build/check times > 5 min works only for the special output
format `-f "time: %e"`. See .makerrc for an example.