use package path instead of package name; closes #15
This commit introduces a new internal variable, PKGNAME, that replaces the old
PKG variable. To find PKGNAME the "Package" field in the PKGDIR/DESCRIPTION
is used. PKG and PKGDIR are now synonyms. The following calls are the same:

  make PKG=cleaver check
  make PKGDIR=cleaver check

The should not be any changes for the user. The behaviour/use of maker is now
more consistent to other R build tools like devtools.